Autobuild High Scores List, Added a Cheat Code
Added a cheat code, when "Cheato" is entered in the main menu, credits
will be equal to one million, however the high score will be marked on
the list. Added an override function highScoresAddPreset and a new
highScoresBuilder function to create a high scores list when there is
none detected. An "Easter Egg" is the initials spell out "BTM Created by
Brian Mascitello for my dad and SoDA ASU". This also showcases the
different possible outputs like "Flush" and how the dates are aligned.
Storing a new cheat variable in the linked list and save files for those
who discover and use the cheat. "Cheato" is a reference to
Banjo-Kazooie, a Nintendo 64 game where a spell book named Cheato aided
to duo with bonuses. Finally, limited the save output to 100 scores.
Even though the game only displays 15, no limit is bad.