Mostly Clean Up and fixing High Score Output
Was reading up about using standard namespace and how it was not
necessarily good practice. Decided to overhaul code with std:: just so
it is all of the right form. Created a helper method
highScoresPrintDate, basically makes highScoresView look nicer and more
readable. Removed unnecessary variables in highScoresLoad and just used
references. It was easier to stick with code at the time, but this is
just cleaner. Also changed the scores list to max length 15, and used a
length modifying for the scores to reduce memory space require to save.