Made a decent Prime Number Solver in Python
Found a weird ratio that works to keep the computational time
reasonable. Could have looked up a premade one but chose to make it on
my own for practice. Not as efficient as many ways, but better than the
layman's function by several times. Sieve of Eratosthenes would be
faster.